일단은 돌아가는 버전 마킹, r2dbc url 쪽 문제로 보임

+++ src/main/kotlin/me/xep/kotlintutorial/model/Product.kt
@@ -1,7 +1,9 @@
 package me.xep.kotlintutorial.model
 
+import org.springframework.data.relational.core.mapping.Table
+
 class Product {
-    var id: Int = 0
+    var id: Long = 0L
     var name: String = ""
     var price: Float = 0.0f
 }
